+++ lnd_methods/unauthenticated/get_wallet_status.d.ts
@@ -2,7 +2,20 @@ import {UnauthenticatedLnd} from '../../lnd_grpc';
 import {UnauthenticatedLightningMethod} from '../../typescript';
 
 export type GetWalletStatusResult = {
-  "state": string;
+  /** Is Wallet Not Created Yet */
+  is_absent?: boolean;
+
+  /** Is Wallet Currently Active */
+  is_active?: boolean;
+
+  /** Is Wallet File Encrypted */
+  is_locked?: boolean;
+
+  /** Is Wallet Starting Up */
+  is_starting?: boolean;
+
+  /** Is Wallet Waiting To Start */
+  is_waiting?: boolean;
 };
 
 /**
@@ -10,7 +23,7 @@ export type GetWalletStatusResult = {
  *
  * Requires `info:read` permission
  */
-export const getState: UnauthenticatedLightningMethod<
+export const getWalletStatus: UnauthenticatedLightningMethod<
   {lnd: UnauthenticatedLnd},
   GetWalletStatusResult
 >;