+++ lnd_responses/index.js
@@ -7,7 +7,6 @@ const failureFromPayment = require('./failure_from_payment');
 const forwardFromHtlcEvent = require('./forward_from_htlc_event');
 const htlcAsPayment = require('./htlc_as_payment');
 const infoAsWalletInfo = require('./info_as_wallet_info');
-const stateAsStateInfo = require('./state_as_state_info');
 const nodeInfoAsNode = require('./node_info_as_node');
 const paymentFailure = require('./payment_failure');
 const paymentRequestDetails = require('./payment_request_details');
@@ -34,6 +33,7 @@ const rpcResolutionAsResolution = require('./rpc_resolution_as_resolution');
 const rpcRouteAsRoute = require('./rpc_route_as_route');
 const rpcTxAsTransaction = require('./rpc_tx_as_transaction');
 const rpcUtxoAsUtxo = require('./rpc_utxo_as_utxo');
+const rpcWalletStateAsState = require('./rpc_wallet_state_as_state');
 
 module.exports = {
   backupsFromSnapshot,
@@ -45,7 +45,6 @@ module.exports = {
   forwardFromHtlcEvent,
   htlcAsPayment,
   infoAsWalletInfo,
-  stateAsStateInfo,
   nodeInfoAsNode,
   paymentFailure,
   paymentRequestDetails,
@@ -72,4 +71,5 @@ module.exports = {
   rpcRouteAsRoute,
   rpcTxAsTransaction,
   rpcUtxoAsUtxo,
+  rpcWalletStateAsState,
 };